+++ UIComponents.Abstractions/Interfaces/FileExplorer/IRelativePath.cs
@@ -1,0 +1,7 @@
+﻿namespace UIComponents.Abstractions.Interfaces.FileExplorer;
+
+public interface IRelativePath
+{
+    public string AbsolutePathReference { get; set; }
+    public string RelativePath { get; set; }
+}