+++ UIComponents.Abstractions/Models/FileExplorer/GetFilesForDirectoryResultModel.cs
@@ -1,0 +1,8 @@
+﻿namespace UIComponents.Abstractions.Models.FileExplorer;
+
+public class GetFilesForDirectoryResultModel
+{
+    public List<UICFileInfo> Files { get; set; } = new();
+    public bool CanCreateInDirectory { get; set; }
+
+}